[WIP] Bracketed paste on windows
[X] Activate bracketed paste only when
ENABLE_VIRTUAL_TERMINAL_PROCESSING is on
[ ] Handle BracketedPasteStart / BracketedPasteEnd
